backend/requirements.txt
fastapi
uvicorn
pyjwt
python-jose[cryptography]
androguard
python-multipart
